requirements.txt
flask
gunicorn
groq
mysql-connector-python
python-dotenv
replicate
